Move `ethernet_*.cpp|h` to `src/network`
