+++ meson.build
@@ -1199,6 +1199,7 @@ subdir('src/dos')
 subdir('src/fpu')
 subdir('src/gui')
 subdir('src/hardware')
+subdir('src/network')
 subdir('src/ints')
 subdir('src/midi')
 subdir('src/shell')
