+++ src/hardware/network/ne2000.cpp
@@ -17,8 +17,8 @@
 #include "hardware/pic.h"
 #include "hardware/port.h"
 #include "hardware/timer.h"
-#include "misc/ethernet.h"
 #include "misc/support.h"
+#include "network/ethernet.h"
 #include "utils/string_utils.h"
 
 /* Couldn't find a real spec for the NE2000 out there, hence this is adapted heavily from Bochs */
