+++ src/misc/meson.build
@@ -3,8 +3,6 @@ libmisc_nomsg_sources = [
     'ansi_code_markup.cpp',
     'console.cpp',
     'cross.cpp',
-    'ethernet.cpp',
-    'ethernet_slirp.cpp',
     'fs_utils.cpp',
     'fs_utils_posix.cpp',
     'fs_utils_win32.cpp',
@@ -31,7 +29,6 @@ libmisc_dependencies = [
     ghc_dep,
     libiir_dep,
     libloguru_dep,
-    libslirp_dep,
     libwhereami_dep,
     sdl2_dep,
     stdcppfs_dep,
