+++ src/network/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(libdosboxcommon PRIVATE 
+  ethernet.cpp
+  ethernet_slirp.cpp
+)
