+++ src/network/ethernet.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText:  2025-2025 The DOSBox Staging Team
 // SPDX-License-Identifier: GPL-2.0-or-later
 
-#include "misc/ethernet.h"
+#include "ethernet.h"
 
 #include <cstring>
